Refactored and addressed PR comments
