+++ domain/src/main/kotlin/es/pedrazamiguez/expenseshareapp/domain/model/GroupPocketBalance.kt
@@ -23,8 +23,10 @@ package es.pedrazamiguez.expenseshareapp.domain.model
  * @param scheduledHoldAmount Sum of groupAmount for SCHEDULED expenses whose dueDate is still
  *                            in the future (in cents). These are not yet spent but reserved.
  *                            When subtracted from virtualBalance, gives the "available" balance.
- * @param totalExtras Sum of all ON_TOP non-discount add-ons across expenses and withdrawals (in cents).
- *                    Includes fees, tips, surcharges on expenses and ATM fees on withdrawals.
+ * @param totalExtras Sum of all non-discount add-ons across expenses (both ON_TOP and INCLUDED)
+ *                    plus ATM fee add-ons on withdrawals (in cents).
+ *                    Includes fees, tips, and surcharges whether they were added on top of the
+ *                    base cost or were already embedded in the user-entered total (INCLUDED).
  *                    Discounts are excluded — they reduce the effective amount but are not
  *                    "extra costs" to surface separately. Always ≥ 0.
  *                    When zero, the UI hides the extras row.
