+++ features/expenses/src/test/kotlin/es/pedrazamiguez/expenseshareapp/features/expense/presentation/viewmodel/handler/ConfigEventHandlerTest.kt
@@ -12,7 +12,8 @@ import es.pedrazamiguez.expenseshareapp.domain.usecase.setting.GetGroupLastUsedC
 import es.pedrazamiguez.expenseshareapp.domain.usecase.setting.GetGroupLastUsedCurrencyUseCase
 import es.pedrazamiguez.expenseshareapp.domain.usecase.setting.GetGroupLastUsedPaymentMethodUseCase
 import es.pedrazamiguez.expenseshareapp.domain.usecase.user.GetMemberProfilesUseCase
-import es.pedrazamiguez.expenseshareapp.features.expense.presentation.mapper.AddExpenseUiMapper
+import es.pedrazamiguez.expenseshareapp.features.expense.presentation.mapper.AddExpenseOptionsMapper
+import es.pedrazamiguez.expenseshareapp.features.expense.presentation.mapper.AddExpenseSplitMapper
 import es.pedrazamiguez.expenseshareapp.features.expense.presentation.viewmodel.action.AddExpenseUiAction
 import es.pedrazamiguez.expenseshareapp.features.expense.presentation.viewmodel.state.AddExpenseUiState
 import io.mockk.coEvery
@@ -87,7 +88,8 @@ class ConfigEventHandlerTest {
             getGroupLastUsedPaymentMethodUseCase = getGroupLastUsedPaymentMethodUseCase,
             getGroupLastUsedCategoryUseCase = getGroupLastUsedCategoryUseCase,
             getMemberProfilesUseCase = getMemberProfilesUseCase,
-            addExpenseUiMapper = AddExpenseUiMapper(localeProvider, resourceProvider),
+            addExpenseOptionsMapper = AddExpenseOptionsMapper(resourceProvider),
+            addExpenseSplitMapper = AddExpenseSplitMapper(localeProvider),
             currencyEventHandler = currencyEventHandler,
             subunitSplitEventHandler = subunitSplitEventHandler
         )
