+++ features/expenses/src/test/kotlin/es/pedrazamiguez/expenseshareapp/features/expense/presentation/viewmodel/handler/SplitEventHandlerTest.kt
@@ -1,12 +1,11 @@
 package es.pedrazamiguez.expenseshareapp.features.expense.presentation.viewmodel.handler
 
 import es.pedrazamiguez.expenseshareapp.core.common.provider.LocaleProvider
-import es.pedrazamiguez.expenseshareapp.core.common.provider.ResourceProvider
 import es.pedrazamiguez.expenseshareapp.core.designsystem.presentation.model.CurrencyUiModel
 import es.pedrazamiguez.expenseshareapp.domain.service.ExpenseCalculatorService
 import es.pedrazamiguez.expenseshareapp.domain.service.split.ExpenseSplitCalculatorFactory
 import es.pedrazamiguez.expenseshareapp.domain.service.split.SplitPreviewService
-import es.pedrazamiguez.expenseshareapp.features.expense.presentation.mapper.AddExpenseUiMapper
+import es.pedrazamiguez.expenseshareapp.features.expense.presentation.mapper.AddExpenseSplitMapper
 import es.pedrazamiguez.expenseshareapp.features.expense.presentation.model.SplitTypeUiModel
 import es.pedrazamiguez.expenseshareapp.features.expense.presentation.model.SplitUiModel
 import es.pedrazamiguez.expenseshareapp.features.expense.presentation.viewmodel.action.AddExpenseUiAction
@@ -80,13 +79,12 @@ class SplitEventHandlerTest {
     @BeforeEach
     fun setUp() {
         val localeProvider = mockk<LocaleProvider>()
-        val resourceProvider = mockk<ResourceProvider>(relaxed = true)
         every { localeProvider.getCurrentLocale() } returns Locale.US
 
         handler = SplitEventHandler(
             splitCalculatorFactory = ExpenseSplitCalculatorFactory(ExpenseCalculatorService()),
             splitPreviewService = SplitPreviewService(),
-            addExpenseUiMapper = AddExpenseUiMapper(localeProvider, resourceProvider)
+            addExpenseUiMapper = AddExpenseSplitMapper(localeProvider)
         )
 
         uiState = MutableStateFlow(baseState)
